api/requirements.txt
connexion == 1.1.15
python_dateutil == 2.6.0
setuptools >= 21.0.0
psycopg2-binary==2.8.2
iptools==0.7.0